backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
boto3==1.35.92
requests==2.32.3
cryptography==44.0.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
PyPDF2==3.0.1
